Add video clip folder
Added a folder to save video clips of tests. Since I can't share clips
over a certain size in Discord, I figured it would be easier to just
attach them to the project. Granted, this will make the project more
than what it needs to do and I should definitely make sure to get rid of
that folder before deploying the game